# Call benchmark method directly