--- a/src/BenchmarkDotNet/Portability/CodeGen.cs
+++ b/src/BenchmarkDotNet/Portability/CodeGen.cs
@@ -7,6 +7,5 @@ namespace BenchmarkDotNet.Portability
     {
         // AggressiveOptimization is not available in netstandard2.0, so just use the value casted to enum.
         public const MethodImplOptions AggressiveOptimizationOption = (MethodImplOptions) 512;
-        public const MethodImplAttributes AggressiveOptimizationOptionForEmit = (MethodImplAttributes) 512;
     }
 }