--- a/src/BenchmarkDotNet/Toolchains/InProcess/Emit/Implementation/Runnable/RunnableConstants.cs
+++ b/src/BenchmarkDotNet/Toolchains/InProcess/Emit/Implementation/Runnable/RunnableConstants.cs
@@ -20,8 +20,6 @@
         public const string GlobalCleanupActionFieldName = "globalCleanupAction";
         public const string IterationSetupActionFieldName = "iterationSetupAction";
         public const string IterationCleanupActionFieldName = "iterationCleanupAction";
-        public const string WorkloadDelegateFieldName = "workloadDelegate";
-        public const string OverheadDelegateFieldName = "overheadDelegate";
         public const string NotElevenFieldName = "NotEleven";
         public const string DummyVarFieldName = "dummyVar";
 
@@ -30,7 +28,9 @@
         public const string Dummy2MethodName = "Dummy2";
         public const string Dummy3MethodName = "Dummy3";
         public const string WorkloadImplementationMethodName = "__Workload";
+        public const string WorkloadWrapperImplementationMethodName = "__WorkloadWrapper";
         public const string OverheadImplementationMethodName = "__Overhead";
+        public const string OverheadWrapperImplementationMethodName = "__OverheadWrapper";
         public const string OverheadActionUnrollMethodName = "OverheadActionUnroll";
         public const string OverheadActionNoUnrollMethodName = "OverheadActionNoUnroll";
         public const string WorkloadActionUnrollMethodName = "WorkloadActionUnroll";