… src/BenchmarkDotNet/Helpers/Reflection.Emit/ModuleBuilderExtensions.cs
… src/BenchmarkDotNet/Templates/BenchmarkType.txt
… tests/BenchmarkDotNet.IntegrationTests.ManualRunning/ExpectedBenchmarkResultsTests.cs