Fixes #1133

This wraps the workload call with a `NoInlining | NoOptimization` method instead of a delegate.

<details><summary>Mac Intel x64 results</summary>
<p>

```
BenchmarkDotNet=v0.13.5.20230619-develop, OS=macOS Monterey 12.3 (21E230) [Darwin 21.4.0]
Intel Core i9-9880H CPU 2.30GHz, 1 CPU, 16 logical and 8 physical cores
.NET SDK=8.0.100-preview.5.23303.2
  [Host]     : .NET 8.0.0 (8.0.23.28008), X64 RyuJIT AVX2
  DefaultJob : .NET 8.0.0 (8.0.23.28008), X64 RyuJIT AVX2
```

Master
|         Method |      Mean |     Error |    StdDev |
|--------------- |----------:|----------:|----------:|
|   OneIncrement | 0.0013 ns | 0.0037 ns | 0.0048 ns |
|   TwoIncrement | 0.0034 ns | 0.0043 ns | 0.0064 ns |
| ThreeIncrement | 0.0000 ns | 0.0000 ns | 0.0000 ns |
|  FourIncrement | 0.1652 ns | 0.0053 ns | 0.0047 ns |
|  FiveIncrement | 0.3950 ns | 0.0047 ns | 0.0039 ns |
|   SixIncrement | 0.6129 ns | 0.0038 ns | 0.0035 ns |

This PR
|         Method |      Mean |     Error |    StdDev |
|--------------- |----------:|----------:|----------:|
|   OneIncrement | 0.2852 ns | 0.0063 ns | 0.0056 ns |
|   TwoIncrement | 0.3822 ns | 0.0477 ns | 0.0742 ns |
| ThreeIncrement | 0.4122 ns | 0.0071 ns | 0.0063 ns |
|  FourIncrement | 0.4845 ns | 0.0080 ns | 0.0071 ns |
|  FiveIncrement | 0.5980 ns | 0.0082 ns | 0.0068 ns |
|   SixIncrement | 1.1828 ns | 0.0078 ns | 0.0069 ns |

</p>
</details>

<details><summary>Windows AMD x64 results</summary>
<p>

```
BenchmarkDotNet=v0.13.5.20230619-develop, OS=Windows 10 (10.0.19045.3086/22H2/2022Update)
AMD Phenom(tm) II X6 1055T Processor, 1 CPU, 6 logical and 6 physical cores
.NET SDK=8.0.100-preview.5.23303.2
  [Host]     : .NET 8.0.0 (8.0.23.28008), X64 RyuJIT SSE3
  DefaultJob : .NET 8.0.0 (8.0.23.28008), X64 RyuJIT SSE3
```

Master
|         Method |      Mean |     Error |    StdDev |
|--------------- |----------:|----------:|----------:|
|   OneIncrement | 0.1833 ns | 0.0359 ns | 0.0300 ns |
|   TwoIncrement | 0.6950 ns | 0.0523 ns | 0.0538 ns |
| ThreeIncrement | 0.7524 ns | 0.0062 ns | 0.0051 ns |
|  FourIncrement | 1.0793 ns | 0.0154 ns | 0.0129 ns |
|  FiveIncrement | 1.1073 ns | 0.0589 ns | 0.0630 ns |
|   SixIncrement | 1.6137 ns | 0.0683 ns | 0.0787 ns |

This PR
|         Method |      Mean |     Error |    StdDev |
|--------------- |----------:|----------:|----------:|
|   OneIncrement | 0.6927 ns | 0.0182 ns | 0.0179 ns |
|   TwoIncrement | 1.1581 ns | 0.0104 ns | 0.0087 ns |
| ThreeIncrement | 1.2368 ns | 0.0057 ns | 0.0050 ns |
|  FourIncrement | 1.4841 ns | 0.0054 ns | 0.0045 ns |
|  FiveIncrement | 1.7893 ns | 0.0918 ns | 0.1128 ns |
|   SixIncrement | 2.3818 ns | 0.1026 ns | 0.0909 ns |

</p>
</details>