@timcassell 2023-07-24
@adamsitnik Is it possible to get `DisassemblyDiagnoser` to work with `InProcessEmitToolchain`?

```
// ! InProcessEmitExecutor, exception: System.InvalidOperationException: Sequence contains no matching element
   at System.Linq.ThrowHelper.ThrowNoMatchException()
   at System.Linq.Enumerable.First[TSource](IEnumerable`1 source, Func`2 predicate)
   at BenchmarkDotNet.Disassemblers.ClrMdV2Disassembler.AttachAndDisassemble(Settings settings) in C:\BenchmarkDotNet-fork\src\BenchmarkDotNet\Disassemblers\ClrMdV2Disassembler.cs:line 40
   at BenchmarkDotNet.Disassemblers.SameArchitectureDisassembler.Disassemble(DiagnoserActionParameters parameters) in C:\BenchmarkDotNet-fork\src\BenchmarkDotNet\Disassemblers\SameArchitectureDisassembler.cs:line 17
   at BenchmarkDotNet.Diagnosers.DisassemblyDiagnoser.Handle(HostSignal signal, DiagnoserActionParameters parameters) in C:\BenchmarkDotNet-fork\src\BenchmarkDotNet\Disassemblers\DisassemblyDiagnoser.cs:line 75
   at BenchmarkDotNet.Diagnosers.CompositeDiagnoser.Handle(HostSignal signal, DiagnoserActionParameters parameters) in C:\BenchmarkDotNet-fork\src\BenchmarkDotNet\Diagnosers\CompositeDiagnoser.cs:line 36
   at BenchmarkDotNet.Toolchains.InProcess.InProcessHost.SendSignal(HostSignal hostSignal) in C:\BenchmarkDotNet-fork\src\BenchmarkDotNet\Toolchains\InProcess\InProcessHost.cs:line 79
   at BenchmarkDotNet.Engines.HostExtensions.AfterAll(IHost host) in C:\BenchmarkDotNet-fork\src\BenchmarkDotNet\Engines\HostExtensions.cs:line 17
   at BenchmarkDotNet.Toolchains.InProcess.Emit.Implementation.RunnableProgram.Run(BenchmarkId benchmarkId, Assembly partitionAssembly, BenchmarkCase benchmarkCase, IHost host) in C:\BenchmarkDotNet-fork\src\BenchmarkDotNet\Toolchains\InProcess\Emit\Implementation\Runnable\RunnableProgram.cs:line 47
   at BenchmarkDotNet.Toolchains.InProcess.Emit.InProcessEmitExecutor.ExecuteCore(IHost host, ExecuteParameters parameters) in C:\BenchmarkDotNet-fork\src\BenchmarkDotNet\Toolchains\InProcess\Emit\InProcessEmitExecutor.cs:line 113
Executable C:\BenchmarkDotNet-fork\tests\BenchmarkDotNet.IntegrationTests.ManualRunning\bin\Release\net7.0\df2e2ee8-ce6c-443b-bfb8-d57ddeb0c04bEmitted.dll not found
```