@timcassell 2023-07-28
I reverted the ClrMd disassembler back to v1 on my local so I could actually inspect the asm. The only difference I see is this.

Default toolchain
```asm
call      qword ptr [BenchmarkDotNet.Autogenerated.Runnable_0.__Overhead()]
```
InProcessEmit
```asm
call      BenchmarkDotNet.Autogenerated.Runnable_0.__Overhead()
```

The IL is exactly the same for those calls, so it seems the JIT treats IL emit slightly different. I don't have any asm knowledge to know what effect that difference makes, but it seems that `qword ptr` is faster. (Only the overhead call is different, the workload call uses `qword ptr` for both toolchains.