@timcassell 2023-08-01
The assembly issue with InProcessEmit is only in net7+ (Core only, Mono is fine). The overhead measurement is off by about 2-3 clock cycles, which isn't far off from the current measurement in all toolchains. I don't think it should block this from being merged.