@timcassell 2025-01-12
@AndyAyersMS I ran the Append_Strings benchmark you mentioned in https://github.com/dotnet/performance/issues/2214#issuecomment-1052585650.

Master:

| Method         | repeat | Mean         | Error      | StdDev     | Allocated |
|--------------- |------- |-------------:|-----------:|-----------:|----------:|
| Append_Strings | 1      |     57.58 ns |   0.317 ns |   0.281 ns |   1.22 KB |
| Append_Strings | 1000   | 12,344.97 ns | 168.074 ns | 157.217 ns | 546.16 KB |

This PR:

| Method         | repeat | Mean         | Error      | StdDev     | Allocated |
|--------------- |------- |-------------:|-----------:|-----------:|----------:|
| Append_Strings | 1      |     60.22 ns |   0.223 ns |   0.186 ns |   1.22 KB |
| Append_Strings | 1000   | 16,915.48 ns | 109.032 ns | 101.988 ns | 546.16 KB |

This PR + #2336 changes:

| Method         | repeat | Mean         | Error     | StdDev    | Allocated |
|--------------- |------- |-------------:|----------:|----------:|----------:|
| Append_Strings | 1      |     57.87 ns |  0.311 ns |  0.275 ns |   1.22 KB |
| Append_Strings | 1000   | 12,625.00 ns | 48.200 ns | 42.728 ns | 546.16 KB |

So it looks like not consuming the return value does fix the GC issue. With these results, I will close this PR and include all changes in #2336, as they clearly need to go together.